refactor: features/post 구조 통합
- /features/post/create, /features/post/[id]를 /features/post로 통합
