--- src/features/(authenticated)/post/create/types/PostCreateForm.types.ts
@@ -1,8 +1,0 @@
-import { TopicType } from '@/features/(authenticated)/post/create/types/Topic.types'
-
-export type PostFormValues = {
-  topic: TopicType
-  title: string
-  content: string
-  draftId?: string // uuid
-}
