+++ src/features/(authenticated)/post/types/Post.types.ts
@@ -1,4 +1,4 @@
-import { TopicType } from '../../create/types/Topic.types'
+import { TopicType } from '../../../../types/Topic.types'
 
 export type Post = {
   writerId: number
@@ -44,3 +44,10 @@ export type CreatePostRequest = {
 }
 
 export type UpdatePostRequest = Omit<CreatePostRequest, 'draftId'> & { postId: number }
+
+export type PostFormValues = {
+  topic: TopicType
+  title: string
+  content: string
+  draftId?: string // uuid
+}
